Bugfix add empty constuctors to DTO and rename guard
