+++ kwetter_angular/src/app/app-routing.module.ts
@@ -3,7 +3,7 @@ import {RouterModule, Routes} from '@angular/router';
 import {UserListComponent} from './component/user-list/user-list.component';
 import {TimelineComponent} from './component/timeline/timeline.component';
 import {LoginComponent} from './component/login/login.component';
-import {AuthenticationGuard} from './service/authentication-guard.service';
+import {AuthenticationGuard} from './guards/authentication.guard';
 import {ProfileComponent} from './component/profile/profile.component';
 import {CreateKwetterComponent} from './component/create-kwetter/create-kwetter.component';
 import {SearchKwetterComponent} from './component/search-kwetter/search-kwetter.component';
@@ -15,7 +15,7 @@ const routes: Routes = [
   {path: 'create-kwetter', component: CreateKwetterComponent, canActivate: [AuthenticationGuard]},
   {path: 'search-kwetter', component: SearchKwetterComponent, canActivate: [AuthenticationGuard]},
   {path: 'login', component: LoginComponent},
-  {path: '**', redirectTo: ''}
+  {path: '**', redirectTo: '/login'}
 ];
 
 @NgModule({
