+++ kwetter_angular/src/app/app.module.ts
@@ -7,7 +7,7 @@ import {AppComponent} from './app.component';
 import {UserListComponent} from './component/user-list/user-list.component';
 import {LoginComponent} from './component/login/login.component';
 import {UserService} from './service/user.service';
-import {AuthenticationGuard} from './service/authentication-guard.service';
+import {AuthenticationGuard} from './guards/authentication.guard';
 import {TimelineComponent} from './component/timeline/timeline.component';
 import {KwetterService} from './service/kwetter.service';
 import {ProfileComponent} from './component/profile/profile.component';
